♻️ rfac(custom): step2, 3 page button 활성화 로직 이동
